refactor(dev): require geom shapes on demand
