+++ src/geom/base.js
@@ -9,7 +9,7 @@ const Util = require('../util');
 const Global = require('../global');
 const Adjust = require('./adjust/index');
 const Labels = require('./label/index');
-const Shape = require('./shape/index');
+const Shape = require('./shape/shape');
 const TooltipMixin = require('./mixin/tooltip');
 const ActiveMixin = require('./mixin/active');
 const SelectMixin = require('./mixin/select');
