+++ src/geom/point.js
@@ -5,6 +5,7 @@
 const View = require('../chart/view');
 const GeomBase = require('./base');
 const Util = require('../util');
+require('./shape/point');
 
 class Point extends GeomBase {
   /**
