+++ src/geom/schema.js
@@ -6,6 +6,7 @@ const View = require('../chart/view');
 const GeomBase = require('./base');
 const Util = require('../util');
 const SizeMixin = require('./mixin/size');
+require('./shape/schema');
 
 class Schema extends GeomBase {
   /**
