--- src/geom/shape/index.js
@@ -1,12 +1,0 @@
-const Shape = require('./shape');
-
-require('./area');
-require('./edge');
-require('./interval');
-require('./line');
-require('./point');
-require('./polygon');
-require('./schema');
-require('./venn');
-
-module.exports = Shape;
