+++ test/unit/geom/shape/edge-spec.js
@@ -1,7 +1,8 @@
 const expect = require('chai').expect;
 const { Canvas } = require('@antv/g');
-const Shape = require('../../../../src/geom/shape/index');
+const Shape = require('../../../../src/geom/shape/shape');
 const Coord = require('../../../../src/coord/');
+require('../../../../src/geom/shape/edge');
 
 const div = document.createElement('div');
 div.id = 'csedge';
